Sound Stuff + Some UI

+++ Magical Girl Metroidvania Jam 2026/Assets/Scripts/UI/StartMenu.cs
@@ -1,6 +1,12 @@
 using UnityEngine;
+using UnityEngine.UI;
 
 public class StartMenu : Menu {
+
+  [SerializeField] private Image logo;
+  [SerializeField] private Button[] uiElements;
+  private bool confirm;
+
   // Start is called once before the first execution of Update after the MonoBehaviour is created
   void Start() {
 
@@ -12,10 +18,60 @@ public class StartMenu : Menu {
   }
 
   public void LoadGame() {
-
+    SaveManager.ptr.BeginningSceneSaveFile();
+    GameManager.ptr.PrepSceneTransition(GameManager.ptr.nextScene);
   }
 
   public void NewGame() {
-    SaveManager.ptr.WipeSave();
+    if (confirm) {
+      SaveManager.ptr.WipeSave();
+
+      eventSystem.firstSelectedGameObject = elementsToSwapTo[2].gameObject;
+
+      UIManager.ptr.NewGame();
+      UIManager.ptr.onFade += IntroSequence;
+
+      for (int i = 0; i < uiElements.Length; i++) {
+        uiElements[i].interactable = false;
+        UIManager.ptr.SetFadeImage(uiElements[i].image, 0.0f);
+      }
+
+      UIManager.ptr.SetFadeImage(logo, 0.0f);
+
+      UIManager.ptr.Fade();
+    } else
+      SwapMenus(1);
+  }
+
+  public void ConfirmNewGame() {
+    confirm = true;
+    NewGame();
+  }
+
+  public void IntroSequence() {
+    gameObject.SetActive(false);
+    GameManager.ptr.p.ToggleControls(true);
+  }
+
+  public void NoLoadedGame() {
+    confirm = true;
+
+    uiElements[0].interactable = false;
+
+    Navigation nav1 = new Navigation();
+    Navigation nav2 = new Navigation();
+
+    nav1.mode = Navigation.Mode.Explicit;
+    nav2.mode = Navigation.Mode.Explicit;
+
+    nav1.selectOnUp = uiElements[3];
+    nav1.selectOnDown = uiElements[2];
+    nav2.selectOnUp = uiElements[2];
+    nav2.selectOnDown = uiElements[1];
+
+    uiElements[1].navigation = nav1;
+    uiElements[3].navigation = nav2;
+
+    eventSystem.SetSelectedGameObject(uiElements[1].gameObject);
   }
 }
